+++ packages/rocketh-deploy-router/CHANGELOG.md
@@ -1,5 +1,12 @@
 # rocketh-deploy-proxy
 
+## 2.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - rocketh-deploy@2.1.0
+
 ## 2.0.2
 
 ### Patch Changes